requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
dvc[s3]
boto3
kafka-python
pytest
ansible
requests
fastparquet
sqlalchemy
fastapi[standard]
pytest-asyncio
httpx
pytest-cov